MNTR-FIX: use only one sequelize instance for the connection, not multiple
